generifying count commands as they make sense for non-decommissioned clusters as well
